[PVR] Remove circular dependency between CPVRGUIActionsChannels and CPVRGUIActionsUtils.
