+++ xbmc/pvr/windows/GUIWindowPVRGuide.cpp
@@ -38,7 +38,6 @@
 #include "pvr/guilib/PVRGUIActionsEPG.h"
 #include "pvr/guilib/PVRGUIActionsPlayback.h"
 #include "pvr/guilib/PVRGUIActionsTimers.h"
-#include "pvr/guilib/PVRGUIActionsUtils.h"
 #include "pvr/recordings/PVRRecordings.h"
 #include "pvr/timers/PVRTimers.h"
 #include "settings/Settings.h"
@@ -79,8 +78,8 @@ void CGUIWindowPVRGuideBase::InitEpgGridControl()
   {
     CPVRManager& mgr = CServiceBroker::GetPVRManager();
 
-    const std::shared_ptr<CPVRChannel> channel =
-        mgr.ChannelGroups()->GetByPath(mgr.Get<PVR::GUI::Utils>().GetSelectedItemPath(m_bRadio));
+    const std::shared_ptr<CPVRChannel> channel = mgr.ChannelGroups()->GetByPath(
+        mgr.Get<PVR::GUI::Channels>().GetSelectedChannelPath(m_bRadio));
 
     if (channel)
     {
@@ -198,7 +197,7 @@ void CGUIWindowPVRGuideBase::UpdateSelectedItemPath()
     const std::shared_ptr<CPVRChannelGroupMember> groupMember =
         epgGridContainer->GetSelectedChannelGroupMember();
     if (groupMember)
-      CServiceBroker::GetPVRManager().Get<PVR::GUI::Utils>().SetSelectedItemPath(
+      CServiceBroker::GetPVRManager().Get<PVR::GUI::Channels>().SetSelectedChannelPath(
           m_bRadio, groupMember->Path());
   }
 }
@@ -229,7 +228,8 @@ bool CGUIWindowPVRGuideBase::Update(const std::string& strDirectory, bool update
     CGUIEPGGridContainer* epgGridContainer = GetGridControl();
     if (epgGridContainer)
       m_bChannelSelectionRestored = epgGridContainer->SetChannel(
-          CServiceBroker::GetPVRManager().Get<PVR::GUI::Utils>().GetSelectedItemPath(m_bRadio));
+          CServiceBroker::GetPVRManager().Get<PVR::GUI::Channels>().GetSelectedChannelPath(
+              m_bRadio));
   }
 
   return bReturn;
